requirements.txt
streamlit
flask
beautifulsoup4
requests
fpdf
python-dotenv
tweepy
schedule
apscheduler
